hotfix: fix of setAlpha for checkboxes
The implementation of the Checkboxes had a serious compatibility error.
The call of setAlpha(float) was introduced in API11. The right call for
API10 is setAlpha(int). Without the fix, the app will crash on Android 2.3
devices.

In addition, the call of setAlpha(int) was deprecated in API17. Another
fix has been implemented in order to be future save.