hotfix-2.0.1: corner radius of buttons fixed
The feature of setting the corner radius of the buttons was not
implemented correctly.

In XML, it was available as float instead of dimension. In code, corner
radius was implemented using a String. This could lead to crashes in
several APIs.

The corner radius has been corrected for XML as well as in code.